Fixed compilation error
